Version 4
Random event gen,

Returns the random exit event of an inputted career.
Will need to get the tables into csv files. One issue is that the adjustments for each event are in a non standard format. An easier solution may be the present the user with the results of the table that need be added. Therefore can just read the entry in the table. In future could standardise the tables myself.